Move packages inside app folder to rspack
