+++ app/aws-lsp-codewhisperer-runtimes/package.json
@@ -8,8 +8,8 @@
         "compile": "tsc --build",
         "copy:wasm": "copyfiles -f --error ../../node_modules/@aws/fully-qualified-names/node/aws_fully_qualified_names_bg.wasm ../../node_modules/web-tree-sitter/tree-sitter.wasm out",
         "copy": "npm run copy:wasm",
-        "package": "npm run compile && npm run copy && npm run webpack",
-        "webpack": "webpack"
+        "package": "npm run compile && npm run copy && npm run rspack",
+        "rspack": "rspack"
     },
     "dependencies": {
         "@aws/language-server-runtimes": "^0.2.20",
@@ -17,8 +17,8 @@
         "copyfiles": "^2.4.1"
     },
     "devDependencies": {
-        "ts-loader": "^9.4.4",
-        "webpack": "^5.94.0",
-        "webpack-cli": "^5.1.4"
+        "@rspack/cli": "^1.0.13",
+        "@rspack/core": "^1.0.13",
+        "ts-loader": "^9.4.4"
     }
 }
