+++ app/aws-lsp-json-runtimes/package.json
@@ -5,16 +5,18 @@
     "main": "out/index.js",
     "scripts": {
         "compile": "tsc --build",
-        "package": "npm run compile && npm run webpack",
+        "package": "npm run compile && npm run rspack",
         "test": "npm run test-integ",
         "test-integ": "npm run package && mocha --timeout 5000 './out/**/*Integ.test.js' --retries 2",
-        "webpack": "webpack"
+        "rspack": "rspack"
     },
     "dependencies": {
         "@aws/language-server-runtimes": "^0.2.20",
         "@aws/lsp-json": "*"
     },
     "devDependencies": {
+        "@rspack/cli": "^1.0.13",
+        "@rspack/core": "^1.0.13",
         "@types/chai": "^4.3.5",
         "@types/chai-as-promised": "^7.1.5",
         "@types/mocha": "^10.0.1",
@@ -22,8 +24,6 @@
         "chai-as-promised": "^7.1.1",
         "mocha": "^10.7.3",
         "ts-loader": "^9.4.4",
-        "ts-lsp-client": "^1.0.3",
-        "webpack": "^5.94.0",
-        "webpack-cli": "^5.1.4"
+        "ts-lsp-client": "^1.0.3"
     }
 }
