+++ app/aws-lsp-partiql-runtimes/package.json
@@ -5,10 +5,9 @@
     "main": "out/index.js",
     "scripts": {
         "clean": "rm -rf out/ bin/ tsconfig.tsbuildinfo",
-        "compile:tsc": "tsc --build",
-        "compile:webpack": "rspack",
-        "compile": "npm run compile:tsc && npm run compile:webpack",
-        "package": "npm run compile && npm run compile:webpack"
+        "compile": "tsc --build && npm run rspack",
+        "package": "npm run rspack",
+        "rspack": "rspack"
     },
     "dependencies": {
         "@aws/language-server-runtimes": "^0.2.20",
@@ -18,8 +17,6 @@
         "@rspack/cli": "^1.0.13",
         "@rspack/core": "^1.0.13",
         "ts-loader": "^9.4.4",
-        "ts-lsp-client": "^1.0.3",
-        "webpack": "^5.94.0",
-        "webpack-cli": "^5.1.4"
+        "ts-lsp-client": "^1.0.3"
     }
 }
