+++ app/aws-lsp-yaml-json-webworker/README.md
@@ -1,6 +1,6 @@
-Package with webpack configuration to bundle LSP Servers with webworker DEXP runtime. 
-Used as an eraly step of LSP Servers integration. 
+Package with rspack configuration to bundle LSP Servers with webworker DEXP runtime. 
+Used as an early step of LSP Servers integration. 
 
 run 
 `npm prepublishOnly` to bundle of LSP server and webworker DEXP runtime.
-`npm serve:webpack` to serve bundled assets locally at 127.0.0.1:5001.
+`npm serve:rspack` to serve bundled assets locally at 127.0.0.1:5001.
