+++ app/aws-lsp-yaml-json-webworker/package.json
@@ -7,8 +7,8 @@
         "clean": "rm -rf dist && rm -rf node_modules && rm -rf build",
         "build": "tsc",
         "prepublishOnly": "npm run build && npm run compile && npm run package",
-        "package": "webpack",
-        "serve:webpack": "NODE_ENV=development webpack serve"
+        "package": "rspack",
+        "serve:rspack": "NODE_ENV=development rspack serve"
     },
     "dependencies": {
         "@aws/language-server-runtimes": "^0.2.20",
@@ -16,14 +16,13 @@
         "@aws/lsp-yaml": "*"
     },
     "devDependencies": {
+        "@rspack/cli": "^1.0.13",
+        "@rspack/core": "^1.0.13",
         "@types/node": "^20.11.30",
         "path": "^0.12.7",
         "path-browserify": "^1.0.1",
         "ts-loader": "^9.5.1",
         "typescript": "^5.5.3",
-        "umd-compat-loader": "^2.1.2",
-        "webpack": "^5.94.0",
-        "webpack-cli": "^5.1.4",
-        "webpack-dev-server": "^5.0.4"
+        "umd-compat-loader": "^2.1.2"
     }
 }
