+++ app/aws-lsp-yaml-runtimes/rspack.config.js
@@ -1,5 +1,4 @@
 var path = require('path')
-var webpack = require('webpack')
 
 const baseConfig = {
     mode: 'development',
