+++ app/hello-world-lsp-runtimes/rspack.config.js
@@ -1,5 +1,5 @@
 var path = require('path')
-var webpack = require('webpack')
+const rspack = require('@rspack/core')
 
 const baseConfig = {
     mode: 'development',
@@ -28,10 +28,26 @@ const baseConfig = {
 const nodeConfig = {
     ...baseConfig,
     entry: {
-        'aws-lsp-json-standalone-with-customization': path.join(__dirname, 'src/serverWithCustomization.ts'),
-        'aws-lsp-json-standalone': path.join(__dirname, 'src/serverWithoutCustomization.ts'),
+        'hello-world-lsp-standalone': path.join(__dirname, 'src/standalone.ts'),
     },
     target: 'node',
 }
 
-module.exports = [nodeConfig]
+const webConfig = {
+    ...baseConfig,
+    entry: {
+        'hello-world-lsp-webworker': path.join(__dirname, 'src/webworker.ts'),
+    },
+    target: 'web',
+    plugins: [
+        new rspack.ProvidePlugin({
+            process: require.resolve('process/browser'),
+        }),
+        new rspack.EnvironmentPlugin({
+            NODE_DEBUG: 'development',
+            READABLE_STREAM: 'disable',
+        }),
+    ],
+}
+
+module.exports = [nodeConfig, webConfig]
